style-fix

+++ resources/views/admin/shortcut/index.blade.php
@@ -3,7 +3,6 @@
 @include('admin.layouts.components.asset_datatables')
 @include('admin.layouts.components.jquery_ui')
 
-
 @section('title')
     <h1>
         Shortcut
@@ -19,14 +18,10 @@
     <div class="box box-info">
         <div class="box-header with-border">
             @if (can('u'))
-                <a href="{{ ci_route('shortcut.form') }}"
-                    class="btn btn-social btn-success btn-sm visible-xs-block visible-sm-inline-block visible-md-inline-block visible-lg-inline-block"><i
-                        class="fa fa-plus"></i> Tambah</a>
+                <a href="{{ ci_route('shortcut.form') }}" class="btn btn-social btn-success btn-sm visible-xs-block visible-sm-inline-block visible-md-inline-block visible-lg-inline-block"><i class="fa fa-plus"></i> Tambah</a>
             @endif
             @if (can('h'))
-                <a href="#confirm-delete" title="Hapus Data"
-                    onclick="deleteAllBox('mainform', '{{ ci_route('shortcut.deleteAll') }}')"
-                    class="btn btn-social btn-danger btn-sm visible-xs-block visible-sm-inline-block visible-md-inline-block visible-lg-inline-block hapus-terpilih"><i
+                <a href="#confirm-delete" title="Hapus Data" onclick="deleteAllBox('mainform', '{{ ci_route('shortcut.deleteAll') }}')" class="btn btn-social btn-danger btn-sm visible-xs-block visible-sm-inline-block visible-md-inline-block visible-lg-inline-block hapus-terpilih"><i
                         class='fa fa-trash-o'></i> Hapus</a>
             @endif
         </div>
